Generate dhparam with turris-cagen instead of using separate dhparam
This changes how dhparam is generated. Until now separate script and
package was used to generate dhparam while turris-cagen is able to do
the same.

This is simple change and allows us to remove one dependency. It takes
some time to generate CA and generating dhparam as part of that makes
sense as well. It is true that it prolongs initial generation (and that
was probably original reason why moving it to separate package) but
this way it is more consistent. It is generated only by user and all is
stored in same location.

Existing configuration is going to be migrated by fix package. That
means missing dhparam is going to be generated and configuration updated
with openvpn reload after that.