Switch from gen_dh to link_dh
The link_dh is new command that allows linking pre-generated dhparam
file instead of generating new one.
There is still an option to generate new one and of course previous one
is not touched. This applies only for new certificates.
There should be no issue in using pre-generated dhparam and at the same
time is is huge time saver.